FILE: moves updateModel
