+++ libs/spirit-islander/util/src/lib/app/update-model.spec.ts
@@ -1,4 +1,4 @@
-import { createSpiritsModel } from '../app/create-model';
+import { createSpiritsModel } from './create-model';
 import { updateModel } from './update-model';
 
 describe('updateModel', () => {
