+++ libs/spirit-islander/util/src/lib/app/update-model.ts
